Fixed argument for 1.8.0

+++ server/on_connect.cpp
@@ -91,7 +91,7 @@ void server::on_read(
 #if UV_VERSION_MAJOR < 1    
     long int nread,
 #else
-    int nread,
+    ssize_t nread,
 #endif
 #if UV_VERSION_MAJOR < 1    
     uv_buf_t buf
